// src/app/models/address-geo.ts
export interface AddressGeo {
    _id?: string;
    name?: string;
    suite?: string;
    streetNumber?: string;
    direction?: string;
    street: string;
    district?: string;
    city?: string;
    provinceCode?: string;
    province?: string;
    postcode?: string;
    countryCode?: string;
    country?: string;
    continentCode?: string;
    continent?: string;
    longitude?: Number;
    latitude?: Number;
    hightitude?: Number;
    description?: string;
}